pull 전 저장
