+++ pet/src/components/mypage/MyPagePost.tsx
@@ -1,12 +1,11 @@
-import React, { SyntheticEvent, useEffect, useState } from 'react';
-import * as S from './style';
+import { SyntheticEvent, useEffect, useState } from 'react';
+import * as S from './MyPagePoststyle';
 import { db } from '../../Firebase';
 import { auth } from '../../Firebase';
 import { collection, getDocs } from 'firebase/firestore';
-import { useNavigate } from 'react-router-dom';
-import defaultImg from '../../assets/images/Caticon1.png'
-
-// import { Link } from 'react-router-dom';
+import { Link } from 'react-router-dom';
+import defaultImg from '../../assets/images/Caticon1.png';
+import NoPosts from './NoPosts';
 
 export type Post = {
   id: string;
@@ -16,12 +15,11 @@ export type Post = {
   createdAt: any;
 };
 
-const MyPagePost = () => {
+const MyPagePost = ({}) => {
   const currentUserInfos = auth.currentUser; // 현재 로그인한 사용자의 정보들(파이어베이스)
   const [posts, setPosts] = useState<Post[]>([]);
   const [isall, setIsall] = useState(false);
   const [loading, setLoading] = useState(true);
-  const navigate = useNavigate();
 
   useEffect(() => {
     const fetchPosts = async () => {
@@ -60,9 +58,6 @@ const MyPagePost = () => {
     isall === false ? setIsall(true) : setIsall(false);
   };
 
-  const navcommunity = () => {
-    navigate('/community');
-  };
 
   const addDefaultImg = (e: SyntheticEvent<HTMLImageElement, Event>) => {
     e.currentTarget.src = defaultImg;
@@ -76,33 +71,40 @@ const MyPagePost = () => {
           <S.ReviewContainer>
             {posts.length < 5
               ? posts.map(post => (
-                  <S.PostContainer key={post.id}>
-                    <S.PostImgContainer src={post.imageUrl} onError={addDefaultImg}/>
-                    <S.TextContainer>
-                      <S.TextTitle> {post.title}</S.TextTitle>
-                      <S.TextIndex> {post.content}</S.TextIndex>
-                    </S.TextContainer>
-                  </S.PostContainer>
-                ))
-              : isall === true
-              ? posts.map(post => (
-                  <S.PostContainer key={post.id}>
-                    <S.PostImgContainer src={post.imageUrl} onError={addDefaultImg}/>
-                    <S.TextContainer>
-                      <S.TextTitle> {post.title}</S.TextTitle>
-                      <S.TextIndex> {post.content}</S.TextIndex>
-                    </S.TextContainer>
-                  </S.PostContainer>
-                ))
-              : posts.map((post, i) =>
-                  i < 4 ? (
+                  <Link key={post.id} to={`/posts/${post.id}`}>
+                    \{' '}
                     <S.PostContainer key={post.id}>
-                      <S.PostImgContainer src={post.imageUrl} onError={addDefaultImg}/>
+                      <S.PostImgContainer src={post.imageUrl} onError={addDefaultImg} />
                       <S.TextContainer>
                         <S.TextTitle> {post.title}</S.TextTitle>
                         <S.TextIndex> {post.content}</S.TextIndex>
                       </S.TextContainer>
                     </S.PostContainer>
+                  </Link>
+                ))
+              : isall === true
+              ? posts.map(post => (
+                  <Link key={post.id} to={`/posts/${post.id}`}>
+                    <S.PostContainer key={post.id}>
+                      <S.PostImgContainer src={post.imageUrl} onError={addDefaultImg} />
+                      <S.TextContainer>
+                        <S.TextTitle> {post.title}</S.TextTitle>
+                        <S.TextIndex> {post.content}</S.TextIndex>
+                      </S.TextContainer>
+                    </S.PostContainer>
+                  </Link>
+                ))
+              : posts.map((post, i) =>
+                  i < 4 ? (
+                    <Link key={post.id} to={`/posts/${post.id}`}>
+                      <S.PostContainer key={post.id}>
+                        <S.PostImgContainer src={post.imageUrl} onError={addDefaultImg} />
+                        <S.TextContainer>
+                          <S.TextTitle> {post.title}</S.TextTitle>
+                          <S.TextIndex> {post.content}</S.TextIndex>
+                        </S.TextContainer>
+                      </S.PostContainer>
+                    </Link>
                   ) : (
                     <></>
                   )
@@ -118,10 +120,8 @@ const MyPagePost = () => {
         </>
       ) : (
         <S.NoPostsContainer>
-          <h2>등록된 게시글이 없습니다</h2>         
-         <S.NoPostsBtn onClick={navcommunity}>게시글 등록하기</S.NoPostsBtn>
-         </S.NoPostsContainer>
-
+          <NoPosts />
+        </S.NoPostsContainer>
       )}
     </S.AllPostContainer>
   );
