+++ pet/src/components/sign-login/LoginPage.tsx
@@ -1,11 +1,13 @@
-import React from 'react';
 import styled from 'styled-components';
 import logoImage from '../../assets/images/logo2.png'; // 로고 이미지 경로
 import Login from './Login'; // 로그인 컴포넌트 경로
 
+
+
 const LoginPage = () => {
+
   return (
-    <Container>
+    <Container >
     <AllContainer>
       <LeftContentContainer>
         <LogoContainer>
@@ -23,23 +25,15 @@ const LoginPage = () => {
 };
 
 const Container = styled.div`
-  @media (max-width: 768px) {
-    margin-left: 4rem;
-  } `
+  max-height: 100vh;
+  overflow: hidden;
+  `
 
 const AllContainer = styled.div`
   display: flex;
   justify-content: center;
   align-items: center;
-  width: 100vw;
   height: 100vh;
-
-  @media (max-width: 768px) {
-    grid-template-columns: 1fr;
-    width: 50rem;
-    height: 80rem;
-    display: flex;
-  }
 `;
 
 const LeftContentContainer = styled.div`
@@ -56,10 +50,6 @@ const RightContentContainer = styled.div`
 const LogoContainer = styled.div`
   width: 100%;
   height: 100%;
-  @media (max-width: 768px) {
-
-}
-
 `;
 
 const LoginContainer = styled.div`
@@ -68,15 +58,11 @@ const LoginContainer = styled.div`
   display: flex;
   justify-content: center;
   align-items: center;
-  @media (max-width: 768px) {
-
-  }
 `;
 
 const LogoImage = styled.img`
   width: 100%;
   height: 100%;
-  /* object-fit: cover; */
 `;
 
 export default LoginPage;
